+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Error handling for adding invoice line from POL. Refs UINV-398.
 * ui-invoice: module warnings analysis. Refs UINV-404.
 * Add copy icon to invoice number. Refs UINV-403.
+* Update label of adjustments field set. Refs UINV-408.
 
 ## [3.1.1](https://github.com/folio-org/ui-invoice/tree/v3.1.1) (2022-03-22)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v3.1.0...v3.1.1)